# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KashFlow API",
    "description": "Achat groupé à prix dégressif. Les paliers portent sur la quantité commandée, jamais sur le nombre de participants (D1), et le prix est rétroactif : quand un palier tombe, tout le groupe en profite (D2).",
    "version": "0.2.0"
  },
  "paths": {
...
    "/orders/{order_id}": {
      "get": {
        "tags": [
          "commandes"
        ],
        "summary": "Get Order",
        "operationId": "get_order_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            }
          },
          "404": {
            "description": "Commande introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}/pay": {
      "post": {
        "tags": [
          "commandes"
        ],
        "summary": "Pay Order",
        "description": "Paiement simulé. Aucun appel externe : le statut passe directement à SUCCESS.\n\nNarratif du pitch : autorisation au join, débit à la clôture (D8).",
        "operationId": "pay_order_orders__order_id__pay_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayOut"
                }
              }
            }
          },
          "404": {
            "description": "Commande introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Commande déjà payée ou annulée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/merchant/products": {
      "get": {
        "tags": [
          "commerçant"
        ],
        "summary": "Lister ses propres offres, brouillons compris",
        "description": "Le catalogue public masque les brouillons ; le commerçant doit les voir.\n\nSans cet endpoint, un produit créé sans grille de paliers devenait\ninvisible à celui-là même qui venait de le saisir.",
        "operationId": "list_my_products_merchant_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantProductsOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "commerçant"
        ],
        "summary": "Créer un produit, avec sa grille de paliers si elle est fournie",
        "operationId": "create_product_merchant_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetail"
                }
              }
            }
          },
          "422": {
            "description": "Grille de paliers incohérente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/merchant/products/{product_id}/tiers": {
      "post": {
        "tags": [
          "commerçant"
        ],
        "summary": "Remplacer la grille de paliers d'un produit",
        "operationId": "set_tiers_merchant_products__product_id__tiers_post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TiersIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TierOut"
                  },
                  "title": "Response Set Tiers Merchant Products  Product Id  Tiers Post"
                }
              }
            }
          },
          "422": {
            "description": "Grille de paliers incohérente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Produit introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Grille figée après paiement ou clôture",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorOut",
        "description": "Réponse d'erreur uniforme : le front ne devine pas."
      },
      "GroupCard": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "share_code": {
            "type": "string",
            "title": "Share Code"
          },
          "status": {
            "$ref": "#/components/schemas/GroupStatus"
          },
          "participants_count": {
            "type": "integer",
            "title": "Participants Count"
          },
          "current_quantity": {
            "type": "integer",
            "title": "Current Quantity"
          },
          "target_quantity": {
            "type": "integer",
            "title": "Target Quantity"
          },
          "current_unit_price": {
            "type": "integer",
            "title": "Current Unit Price"
          },
          "progress_ratio": {
            "type": "number",
            "title": "Progress Ratio"
          },
          "seconds_remaining": {
            "type": "integer",
            "title": "Seconds Remaining"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "share_code",
          "status",
          "participants_count",
          "current_quantity",
          "target_quantity",
          "current_unit_price",
          "progress_ratio",
          "seconds_remaining"
        ],
        "title": "GroupCard"
      },
...
      "GroupStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "LOCKED",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "GroupStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MerchantProductRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "unit_label": {
            "type": "string",
            "title": "Unit Label"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "stock": {
            "type": "integer",
            "title": "Stock"
          },
          "individual_price": {
            "type": "integer",
            "title": "Individual Price"
          },
          "best_price": {
            "type": "integer",
            "title": "Best Price"
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatus"
          },
          "tiers": {
            "items": {
              "$ref": "#/components/schemas/TierOut"
            },
            "type": "array",
            "title": "Tiers"
          },
          "groups_count": {
            "type": "integer",
            "title": "Groups Count"
          },
          "reserved_units": {
            "type": "integer",
            "title": "Reserved Units"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "unit_label",
          "image_url",
          "stock",
          "individual_price",
          "best_price",
          "status",
          "tiers",
          "groups_count",
          "reserved_units"
        ],
        "title": "MerchantProductRow",
        "description": "Une offre vue par son commerçant, grille comprise.\n\nLe catalogue public (`ProductCard`) masque les brouillons et n'expose pas le\nstock : le commerçant, lui, doit voir ses deux."
      },
      "MerchantProductsOut": {
        "properties": {
          "products": {
            "items": {
              "$ref": "#/components/schemas/MerchantProductRow"
            },
            "type": "array",
            "title": "Products"
          }
        },
        "type": "object",
        "required": [
          "products"
        ],
        "title": "MerchantProductsOut"
      },
...
      "OrderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "group_id": {
            "type": "integer",
            "title": "Group Id"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "unit_label": {
            "type": "string",
            "title": "Unit Label"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "integer",
            "title": "Unit Price"
          },
          "total_amount": {
            "type": "integer",
            "title": "Total Amount"
          },
          "individual_price": {
            "type": "integer",
            "title": "Individual Price"
          },
          "saving": {
            "type": "integer",
            "title": "Saving"
          },
          "payment_status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "order_status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "group_id",
          "group_name",
          "product_id",
          "product_name",
          "unit_label",
          "quantity",
          "unit_price",
          "total_amount",
          "individual_price",
          "saving",
          "payment_status",
          "order_status",
          "created_at"
        ],
        "title": "OrderOut"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONFIRMED",
          "CANCELLED"
        ],
        "title": "OrderStatus"
      },
      "PayOut": {
        "properties": {
          "payment": {
            "$ref": "#/components/schemas/PaymentOut"
          },
          "order": {
            "$ref": "#/components/schemas/OrderOut"
          }
        },
        "type": "object",
        "required": [
          "payment",
          "order"
        ],
        "title": "PayOut"
      },
      "PaymentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "transaction_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Reference"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "amount",
          "method",
          "status",
          "transaction_reference"
        ],
        "title": "PaymentOut"
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "SUCCESS",
          "FAILED",
          "REFUNDED"
        ],
        "title": "PaymentStatus"
      },
...
      "ProductCreateIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "unit_label": {
            "type": "string",
            "maxLength": 20,
            "title": "Unit Label",
            "default": "unité"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "stock": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Stock"
          },
          "individual_price": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Individual Price"
          },
          "tiers": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TierIn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tiers"
          }
        },
        "type": "object",
        "required": [
          "name",
          "stock",
          "individual_price"
        ],
        "title": "ProductCreateIn"
      },
      "ProductDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "unit_label": {
            "type": "string",
            "title": "Unit Label"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "stock": {
            "type": "integer",
            "title": "Stock"
          },
          "individual_price": {
            "type": "integer",
            "title": "Individual Price"
          },
          "merchant_name": {
            "type": "string",
            "title": "Merchant Name"
          },
          "merchant_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Location"
          },
          "tiers": {
            "items": {
              "$ref": "#/components/schemas/TierOut"
            },
            "type": "array",
            "title": "Tiers"
          },
          "open_groups": {
            "items": {
              "$ref": "#/components/schemas/GroupCard"
            },
            "type": "array",
            "title": "Open Groups"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "unit_label",
          "image_url",
          "stock",
          "individual_price",
          "merchant_name",
          "merchant_location",
          "tiers",
          "open_groups"
        ],
        "title": "ProductDetail"
      },
      "ProductStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "ACTIVE",
          "CLOSED"
        ],
        "title": "ProductStatus"
      },
...
      "TierIn": {
        "properties": {
          "min_quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Min Quantity"
          },
          "max_quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Quantity"
          },
          "unit_price": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Unit Price"
          }
        },
        "type": "object",
        "required": [
          "min_quantity",
          "unit_price"
        ],
        "title": "TierIn"
      },
      "TierOut": {
        "properties": {
          "min_quantity": {
            "type": "integer",
            "title": "Min Quantity"
          },
          "unit_price": {
            "type": "integer",
            "title": "Unit Price"
          }
        },
        "type": "object",
        "required": [
          "min_quantity",
          "unit_price"
        ],
        "title": "TierOut"
      },
...
      "TiersIn": {
        "properties": {
          "tiers": {
            "items": {
              "$ref": "#/components/schemas/TierIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Tiers"
          }
        },
        "type": "object",
        "required": [
          "tiers"
        ],
        "title": "TiersIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}